+++ sdk/src/test/java/com/hedera/hashgraph/sdk/TransactionIdTest.java
@@ -187,4 +187,11 @@ class TransactionIdTest {
             () -> TransactionId.fromString("0.0.23847@1588539964/4")
         );
     }
+
+    @Test
+    void shouldAddTrailingZeroesToNanoseconds() {
+        var txIdString = "0.0.4163533@1681876267.054802581";
+        var txId = TransactionId.fromString(txIdString);
+        assertThat(txId).hasToString(txIdString);
+    }
 }