+++ CHANGELOG.md
@@ -8,7 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Fixed
+ - The `transactionId` nanoseconds are now left padded with 0s for 9 chars
  - Importing and exporting ECDSA keys to DER format is now compatible with OpenSSL
+
+## 2.24.0
+
+### Added
+ - Alias support in `AccountCreateTransaction`
+ - `CreateAccountWithAliasExample`
+ - `CreateAccountWithAliasAndReceiverSignatureRequiredExample`
 
 ## 2.22.0